moved telemetry package
