+++ tm2/pkg/telemetry/init.go
